requirements.txt
audiosearch==1.0.3
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
Pygments==2.2.0
pytest==3.2.3
pytest-mock==1.6.3
PyYAML==3.12
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
